accounts/urls.py
from django.urls import path
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView
from .views import UserRegistrationView, UserProfileView # New import for the profile view

urlpatterns = [
    path('register/', UserRegistrationView.as_view(), name='register'),
    
    # JWT Login Endpoints
    path('login/', TokenObtainPairView.as_view(), name='login'),
    path('login/refresh/', TokenRefreshView.as_view(), name='token_refresh'),

    # The new protected route for user profiles
    path('profile/', UserProfileView.as_view(), name='profile'),
]